requirements.txt
django==2.2.2
django-cors-headers==3.0.2
django-rest-auth==0.9.5
django-tenants==2.2.3
djangorestframework==3.9.4
pillow==6.0.0
psycopg2==2.8.2
psycopg2-binary==2.8.2
pytz==2019.1
sqlparse==0.3.0
reportlab==3.5.23
xhtml2pdf==0.2.3
